double to float

+++ Tinn.c
@@ -6,55 +6,55 @@
 #include <time.h>
 
 // Error function.
-static double err(double a, double b)
+static float err(float a, float b)
 {
-    return 0.5 * pow(a - b, 2.0);
+    return 0.5f * powf(a - b, 2.0f);
 }
 
 // Partial derivative of error function.
-static double pderr(double a, double b)
+static float pderr(float a, float b)
 {
     return a - b;
 }
 
 // Total error.
-static double terr(const double* tg, const double* o, int size)
+static float terr(const float* tg, const float* o, int size)
 {
-    double sum = 0.0;
+    float sum = 0.0f;
     for(int i = 0; i < size; i++)
         sum += err(tg[i], o[i]);
     return sum;
 }
 
 // Activation function.
-static double act(double a)
+static float act(float a)
 {
-    return 1.0 / (1.0 + exp(-a));
+    return 1.0f / (1.0f + expf(-a));
 }
 
 // Partial derivative of activation function.
-static double pdact(double a)
+static float pdact(float a)
 {
-    return a * (1.0 - a);
+    return a * (1.0f - a);
 }
 
 // Floating point random from 0.0 - 1.0.
-static double frand()
+static float frand()
 {
-    return rand() / (double) RAND_MAX;
+    return rand() / (float) RAND_MAX;
 }
 
 // Back propagation.
-static void backwards(const Tinn t, const double* in, const double* tg, double rate)
+static void backwards(const Tinn t, const float* in, const float* tg, float rate)
 {
     for(int i = 0; i < t.nhid; i++)
     {
-        double sum = 0.0;
+        float sum = 0.0f;
         // Calculate total error change with respect to output.
         for(int j = 0; j < t.nops; j++)
         {
-            double a = pderr(t.o[j], tg[j]);
-            double b = pdact(t.o[j]);
+            float a = pderr(t.o[j], tg[j]);
+            float b = pdact(t.o[j]);
             sum += a * b * t.x[j * t.nhid + i];
             // Correct weights in hidden to output layer.
             t.x[j * t.nhid + i] -= rate * a * b * t.h[i];
@@ -66,12 +66,12 @@ static void backwards(const Tinn t, const double* in, const double* tg, double r
 }
 
 // Forward propagation.
-static void forewards(const Tinn t, const double* in)
+static void forewards(const Tinn t, const float* in)
 {
     // Calculate hidden layer neuron values.
     for(int i = 0; i < t.nhid; i++)
     {
-        double sum = 0.0;
+        float sum = 0.0f;
         for(int j = 0; j < t.nips; j++)
             sum += in[j] * t.w[i * t.nips + j];
         t.h[i] = act(sum + t.b[0]);
@@ -79,7 +79,7 @@ static void forewards(const Tinn t, const double* in)
     // Calculate output layer neuron values.
     for(int i = 0; i < t.nops; i++)
     {
-        double sum = 0.0;
+        float sum = 0.0f;
         for(int j = 0; j < t.nhid; j++)
             sum += t.h[j] * t.x[i * t.nhid + j];
         t.o[i] = act(sum + t.b[1]);
@@ -89,17 +89,17 @@ static void forewards(const Tinn t, const double* in)
 // Randomizes weights and biases.
 static void twrand(const Tinn t)
 {
-    for(int i = 0; i < t.nw; i++) t.w[i] = frand() - 0.5;
-    for(int i = 0; i < t.nb; i++) t.b[i] = frand() - 0.5;
+    for(int i = 0; i < t.nw; i++) t.w[i] = frand() - 0.5f;
+    for(int i = 0; i < t.nb; i++) t.b[i] = frand() - 0.5f;
 }
 
-double* xpredict(const Tinn t, const double* in)
+float* xpredict(const Tinn t, const float* in)
 {
     forewards(t, in);
     return t.o;
 }
 
-double xttrain(const Tinn t, const double* in, const double* tg, double rate)
+float xttrain(const Tinn t, const float* in, const float* tg, float rate)
 {
     forewards(t, in);
     backwards(t, in, tg, rate);
@@ -112,11 +112,11 @@ Tinn xtbuild(int nips, int nhid, int nops)
     // Tinn only supports one hidden layer so there are two biases.
     t.nb = 2;
     t.nw = nhid * (nips + nops);
-    t.w = (double*) calloc(t.nw, sizeof(*t.w));
+    t.w = (float*) calloc(t.nw, sizeof(*t.w));
     t.x = t.w + nhid * nips;
-    t.b = (double*) calloc(t.nb, sizeof(*t.b));
-    t.h = (double*) calloc(nhid, sizeof(*t.h));
-    t.o = (double*) calloc(nops, sizeof(*t.o));
+    t.b = (float*) calloc(t.nb, sizeof(*t.b));
+    t.h = (float*) calloc(nhid, sizeof(*t.h));
+    t.o = (float*) calloc(nops, sizeof(*t.o));
     t.nips = nips;
     t.nhid = nhid;
     t.nops = nops;
@@ -131,8 +131,8 @@ void xtsave(const Tinn t, const char* path)
     // Header.
     fprintf(file, "%d %d %d\n", t.nips, t.nhid, t.nops);
     // Biases and weights.
-    for(int i = 0; i < t.nb; i++) fprintf(file, "%lf\n", t.b[i]);
-    for(int i = 0; i < t.nw; i++) fprintf(file, "%lf\n", t.w[i]);
+    for(int i = 0; i < t.nb; i++) fprintf(file, "%f\n", (double) t.b[i]);
+    for(int i = 0; i < t.nw; i++) fprintf(file, "%f\n", (double) t.w[i]);
     fclose(file);
 }
 
@@ -147,8 +147,8 @@ Tinn xtload(const char* path)
     // A new tinn is returned.
     Tinn t = xtbuild(nips, nhid, nips);
     // Biases and weights.
-    for(int i = 0; i < t.nb; i++) fscanf(file, "%lf\n", &t.b[i]);
-    for(int i = 0; i < t.nw; i++) fscanf(file, "%lf\n", &t.w[i]);
+    for(int i = 0; i < t.nb; i++) fscanf(file, "%f\n", &t.b[i]);
+    for(int i = 0; i < t.nw; i++) fscanf(file, "%f\n", &t.w[i]);
     fclose(file);
     return t;
 }
